# swagger.yaml
openapi: 3.0.3
info:
  title: Marketplace Management System API
  description: RESTful API documentation for modern marketplace management system
  version: 1.0.0
  contact:
    name: Marketplace Management System
    email: info@marketplace.com
  license:
    name: MIT License
    url: https://opensource.org/licenses/MIT
servers:
  - url: http://localhost:8080
    description: Development Server
  - url: https://api.marketplace.com
    description: Production Server
...
paths:
  /api/auth/login:
    post:
      tags:
        - Authentication
      summary: User Login
      description: Login with email and password
      requestBody:
        required: true
        content:
          application/json:
            schema:
              $ref: '#/components/schemas/LoginRequest'
      responses:
        '200':
          description: Başarılı giriş
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/LoginResponse'
        '401':
          description: Giriş başarısız
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/LoginResponse'
        '500':
          description: Sunucu hatası
  /api/auth/register:
    post:
      tags:
        - Authentication
      summary: User Registration
      description: Create new user account
      requestBody:
        required: true
        content:
          application/json:
            schema:
              $ref: '#/components/schemas/User'
      responses:
        '200':
          description: Başarılı kayıt
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/LoginResponse'
        '400':
          description: Geçersiz istek
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/LoginResponse'
        '500':
          description: Sunucu hatası
  /api/auth/refresh:
    post:
      tags:
        - Authentication
      summary: Token yenileme
      description: Get new access token using refresh token
      requestBody:
        required: true
        content:
          application/json:
            schema:
              $ref: '#/components/schemas/RefreshTokenRequest'
      responses:
        '200':
          description: Token başarıyla yenilendi
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/LoginResponse'
        '401':
          description: Geçersiz refresh token
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/LoginResponse'
        '500':
          description: Sunucu hatası
...
  /api/products:
    get:
      tags:
        - Products
      summary: Tüm ürünleri getir
      responses:
        '200':
          description: Başarılı
          content:
            application/json:
              schema:
                type: array
                items:
                  $ref: '#/components/schemas/Product'
    post:
      tags:
        - Products
      summary: Yeni ürün oluştur
      requestBody:
        required: true
        content:
          application/json:
            schema:
              $ref: '#/components/schemas/Product'
      responses:
        '200':
          description: Ürün oluşturuldu
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/Product'
  /api/products/{id}:
    get:
      tags:
        - Products
      summary: ID ile ürün getir
      parameters:
        - name: id
          in: path
          required: true
          schema:
            type: integer
      responses:
        '200':
          description: Başarılı
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/Product'
        '404':
          description: Ürün bulunamadı
    put:
      tags:
        - Products
      summary: Ürün güncelle
      parameters:
        - name: id
          in: path
          required: true
          schema:
            type: integer
      requestBody:
        required: true
        content:
          application/json:
            schema:
              $ref: '#/components/schemas/Product'
      responses:
        '200':
          description: Ürün güncellendi
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/Product'
        '404':
          description: Ürün bulunamadı
    delete:
      tags:
        - Products
      summary: Ürün sil
      parameters:
        - name: id
          in: path
          required: true
          schema:
            type: integer
      responses:
        '200':
          description: Ürün silindi
        '404':
          description: Ürün bulunamadı
  /api/marketplaces:
    get:
      tags:
        - Marketplaces
      summary: Tüm pazarları getir
      responses:
        '200':
          description: Başarılı
          content:
            application/json:
              schema:
                type: array
                items:
                  $ref: '#/components/schemas/MarketplaceResponse'
    post:
      tags:
        - Marketplaces
      summary: Yeni pazar oluştur
      requestBody:
        required: true
        content:
          application/json:
            schema:
              $ref: '#/components/schemas/Marketplace'
      responses:
        '200':
          description: Pazar oluşturuldu
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/Marketplace'
  /api/marketplaces/{id}:
    get:
      tags:
        - Marketplaces
      summary: ID ile pazar getir
      parameters:
        - name: id
          in: path
          required: true
          schema:
            type: integer
      responses:
        '200':
          description: Başarılı
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/MarketplaceResponse'
        '404':
          description: Pazar bulunamadı
    put:
      tags:
        - Marketplaces
      summary: Pazar güncelle
      parameters:
        - name: id
          in: path
          required: true
          schema:
            type: integer
      requestBody:
        required: true
        content:
          application/json:
            schema:
              $ref: '#/components/schemas/Marketplace'
      responses:
        '200':
          description: Pazar güncellendi
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/Marketplace'
        '404':
          description: Pazar bulunamadı
    delete:
      tags:
        - Marketplaces
      summary: Pazar sil
      parameters:
        - name: id
          in: path
          required: true
          schema:
            type: integer
      responses:
        '200':
          description: Pazar silindi
        '404':
          description: Pazar bulunamadı
...
components:
  schemas:
    LoginRequest:
      type: object
      required:
        - email
        - password
      properties:
        email:
          type: string
          format: email
          example: user@example.com
        password:
          type: string
          format: password
          example: password123
    LoginResponse:
      type: object
      properties:
        id:
          type: integer
          format: int64
        email:
          type: string
        role:
          $ref: '#/components/schemas/Role'
        name:
          type: string
        phoneNumber:
          type: string
        token:
          type: string
        refreshToken:
          type: string
        expiresIn:
          type: integer
          format: int64
        success:
          type: boolean
    RefreshTokenRequest:
      type: object
      required:
        - refreshToken
      properties:
        refreshToken:
          type: string
    User:
      type: object
      properties:
        id:
          type: integer
          format: int64
        email:
          type: string
          format: email
        password:
          type: string
          format: password
        role:
          $ref: '#/components/schemas/Role'
        name:
          type: string
        phoneNumber:
          type: string
    Role:
      type: string
      enum:
        - BELEDIYE
        - ZABITA
        - VENDOR
        - CUSTOMER
    Product:
      type: object
      properties:
        id:
          type: integer
          format: int64
        name:
          type: string
        category:
          type: string
        unit:
          type: string
        purchasePrice:
          type: number
          format: double
        maxSalePrice:
          type: number
          format: double
        currentSalePrice:
          type: number
          format: double
        stock:
          type: integer
        isAvailable:
          type: boolean
        imageUrl:
          type: string
        description:
          type: string
        createdAt:
          type: string
          format: date-time
        updatedAt:
          type: string
          format: date-time
    Marketplace:
      type: object
      properties:
        id:
          type: integer
          format: int64
        name:
          type: string
        latitude:
          type: number
          format: double
        longitude:
          type: number
          format: double
        address:
          type: string
        isOpenToday:
          type: boolean
        openDays:
          type: string
        openTime:
          type: string
        closeTime:
          type: string
        imageUrl:
          type: string
        schemaData:
          type: string
    MarketplaceResponse:
      type: object
      properties:
        id:
          type: integer
          format: int64
        name:
          type: string
        latitude:
          type: number
          format: double
        longitude:
          type: number
          format: double
        address:
          type: string
        isOpenToday:
          type: boolean
        openDays:
          type: array
          items:
            type: string
        openTime:
          type: string
        closeTime:
          type: string
        imageUrl:
          type: string
        schemaData:
          type: string
...